Bounding Volume library work
The BV_Lib is getting a major overhaul. I created a base class and used
the new c++ 11 keyword final on the derived classes so hopfully this
will let the compilder make the bounding volumes just as fast as a non
virtual class.
Some functions are not completed yet for the OBB.