New UI Library
I am beginning work on a new UI library. This library is called, wait
for it.... MY_UI_Too. I dont put much emphasis on the name of the things
I create.
This library is going to be more interface and contract driven than the
previous iteration. This iteration is really going to be great. It
doesnt actually do anything yet. Currently, I am working on the skin,
which should be done shortly.